Add function to check if the current player has won, and make bug fixes
The checkWin() function combines functions defined previously as helper functions to check for all kinds of wins in one place. The bug fixes consist of changing the notFinished() function to the boardIsFilled() function, as the game's result couldn't be evaluated until the board was filled up entirely. Additionally, in the checkWin() helper functions, empty cells were accidentally comparable.